+++ src/main/frontend/src/components/Auth.tsx
@@ -1,22 +1,14 @@
 import {createContext, useContext, useState} from "react";
-import {Navigate, Outlet} from "react-router-dom";
-
-
-interface AuthContextType {
-    user: Login | undefined;
-    signIn: (data: any) => void;
-    signOut: () => void;
-}
 
 const AuthContext = createContext<AuthContextType | undefined>(undefined);
 
 function AuthProvider({ children}: Props) {
-    const [user, setUser] = useState<any | undefined>(() => {
+    const [user, setUser] = useState<Login | undefined>(() => {
         const user = sessionStorage.getItem("user");
         return user ? JSON.parse(user) : undefined;
     });
 
-    const signIn = (data: any) => {
+    const signIn = (data: Login) => {
         setUser(data);
         sessionStorage.setItem("user", JSON.stringify(data));
     };
@@ -41,13 +33,4 @@ const useAuth = () => {
     return context;
 }
 
-function ProtectedRoute() {
-    const {user} = useAuth();
-    console.log(`User2: ${user}`);
-    if (user === undefined) {
-        return <Navigate to={"/login"} replace />;
-    }
-    return <Outlet />;
-}
-
-export { AuthProvider, useAuth, ProtectedRoute };
+export { AuthProvider, useAuth };
